backend/app/core/config.py
import os

from dotenv import load_dotenv

load_dotenv()

GROQ_API_KEY = os.getenv("GROQ_API_KEY", "")
GROQ_PRIMARY_MODEL = os.getenv("GROQ_PRIMARY_MODEL", "llama-3.1-8b-instant")
GROQ_ROUTING_MODEL = os.getenv("GROQ_ROUTING_MODEL", "llama-3.3-70b-versatile")
GROQ_FALLBACK_MODEL = os.getenv("GROQ_FALLBACK_MODEL", "llama-3.3-70b-versatile")
DATABASE_URL = os.getenv(
	"DATABASE_URL",
	"mysql+pymysql://root:root123@localhost:3306/ai_crm",
)
FRONTEND_ORIGIN = os.getenv("FRONTEND_ORIGIN", "http://localhost:3000")