backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
python-dateutil
groq
langgraph
cryptography